Restructure project into packages for tasks, parser, storage, ui
