+++ src/main/java/john/John.java
@@ -1,5 +1,11 @@
 package john;
 
+import john.command.Parser;
+import john.data.Storage;
+import john.data.TaskList;
+import john.exceptions.JohnException;
+import john.tasks.Task;
+
 import java.io.IOException;
 import java.nio.file.Path;
 import java.util.List;
