+++ src/main/java/john/tasks/Task.java
@@ -1,4 +1,4 @@
-package john;
+package john.tasks;
 
 /**
  * A Task class has the attributes description and isDone, where the former is a string that describes a given task
